requirements.txt
"""
python-telegram-bot==20.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
fpdf2==2.5.5
qrcode==7.4
Pillow==9.5.0
arabic-reshaper==0.3.1
python-bidi==0.4.2
uuid
"""
